Moved CartPoleEnv class
