+++ gym_ignition/__init__.py
@@ -8,7 +8,7 @@ from gym_ignition.envs.base.ignition_python_env import IgnitionPythonEnv
 
 register(
     id='CartPoleIgnition-v0',
-    entry_point='gym_ignition.envs.cartpole:CartPoleEnv')
+    entry_point='gym_ignition.envs.cpp.cartpole:CartPoleEnv')
 
 register(
     id='CartPoleIgnitionPython-v0',
