+++ gym_ignition/envs/cpp/__init__.py
@@ -2,4 +2,4 @@
 # This software may be modified and distributed under the terms of the
 # GNU Lesser General Public License v2.1 or any later version.
 
-from gym_ignition.envs.cartpole.CartPole import CartPoleEnv
+from gym_ignition.envs.cpp.cartpole import CartPoleEnv
